refactor: separated context and provider
