+++ src/routes/routes.tsx
@@ -2,10 +2,10 @@ import { NavigationContainer } from '@react-navigation/native';
 import { AuthStack } from './auth-stack';
 import { AppStack } from './app-stack';
 import { useContext } from 'react';
-import { AuthContext } from '../contexts/auth';
+import { AuthContext } from '../contexts/auth-context';
 
 export function Routes() {
-  const authContext = useContext(AuthContext)
+  const authContext = useContext(AuthContext);
 
   return (
     <>
